LayoutTests/inspector/debugger/script-formatter-expected.txt

Tests the script formatting functionality.


Running: testScriptFormatterWorker
